+++ helper/date.js
@@ -1,16 +1,16 @@
-function setDate (params){
-    let date ;
-    if(params){
-         date = new Date(params)
-    }
-    else {
-         date = new Date()
-    }
-    let dd = date.getDate()
-    let mm = date.getMonth()
-    let yy = date.getFullYear()
-    let today =  `${dd}-${mm}-${yy}`
-    return today
-}
+// function setDate (params){
+//     let date ;
+//     if(params){
+//          date = new Date(params)
+//     }
+//     else {
+//          date = new Date()
+//     }
+//     let dd = date.getDate()
+//     let mm = date.getMonth()
+//     let yy = date.getFullYear()
+//     let today =  `${dd}-${mm}-${yy}`
+//     return today
+// }
 
-module.exports = setDate
+// module.exports = setDate
